src/test/kotlin/org/carth/aoc20/Day05Tests.kt
package org.carth.aoc20

import org.carth.common.DayTests
import org.junit.jupiter.api.*
import org.junit.jupiter.api.Assertions.assertEquals


class Day05Tests : DayTests() {

    @Nested
    @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
    inner class Day05PartOneTest {

        @Test
        @Order(1)
        fun `solve part one of sample`() {
            // Act
            val answer = Day05(testInputAsListOfString()).solvePartOne()

            // Assert
            assertEquals( 820, answer)
        }

        @Test
        @Order(2)
        fun `solve part one`() {
            // Act
            val answer = Day05(inputAsListOfString()).solvePartOne()

            // Assert
            assertEquals( 801, answer)
        }
    }

    @Nested
    @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
    inner class Day05PartTwoTest {

        @Test
        @Order(1)
        fun `solve part two of sample`() {
            // Act
            val answer = Day05(testInputAsListOfString()).solvePartTwo()

            // Assert
            assertEquals( 120, answer)
        }

        @Test
        @Order(2)
        fun `solve part two`() {
            // Act
            val answer = Day05(inputAsListOfString()).solvePartTwo()
            // Assert
            assertEquals( 597, answer)
        }
    }
}